requirements.txt
streamlit
google-generativeai
pandas
scikit-learn
numpy
python-dotenv
